+++ app/components/Editor.vue
@@ -227,7 +227,7 @@ import { ref, watch, toRefs, computed, onMounted, onUnmounted } from 'vue';
 import Fuse from 'fuse.js';
 import groupedEmojis from '~/data/emojis';
 import { useResizeObserver } from '@vueuse/core';
-import { debounce, orderBy, flatten } from 'lodash';
+import { debounce, flatten } from 'lodash';
 import useI18n from '@/composables/useI18n';
 
 // @see https://github.com/muan/unicode-emoji-json
@@ -268,6 +268,7 @@ defineEmits([
 const { sizes, orientation, language } = toRefs(props);
 const { $bus, $shiki } = useNuxtApp();
 const { t } = useI18n();
+const { options: languageOptions } = useLanguages();
 
 const width = ref(0);
 const height = ref(0);
@@ -279,7 +280,7 @@ const filteredEmojis = ref(emojis);
 
 const fuse = new Fuse(emojis, { keys: ['name'] });
 
-const languages = computed(() => orderBy($shiki.languages()));
+const languages = computed(() => languageOptions($shiki.languages()));
 const landscape = computed(() => ['left', 'right'].includes(orientation.value));
 
 const languageAlias = computed(