+++ app/components/ModalPreferences.vue
@@ -493,6 +493,7 @@ defineProps({
 defineEmits(['update:modelValue']);
 
 const { $shiki } = useNuxtApp();
+const { options: languageOptions } = useLanguages();
 const activeTab = ref('editor');
 const isAutoColorScheme = ref(null);
 const preferences = usePreferencesStore();
@@ -501,7 +502,7 @@ const { types: socialTypes, positions: socialPositions } = useSocials();
 
 const { fontFamilies } = useFonts();
 
-const languages = computed(() => orderBy($shiki.languages()));
+const languages = computed(() => languageOptions($shiki.languages()));
 
 const tabs = computed(() => [
     { value: 'editor', label: t('preferences.editor'), icon: CodeIcon },